+++ R/engine.R
@@ -35,6 +35,13 @@ knit_engines = new_defaults()
 #'
 #' If you have designed a language engine, you may call this function in the end
 #' to format and return the text output from your engine.
+#'
+#' For expert users, an advanced usage of this function is
+#' \code{engine_output(options, out = LIST)} where \code{LIST} is a list that
+#' has the same structure as the output of \code{evaluate::evaluate()}. In this
+#' case, the arguments \code{code} and \code{extra} are ignored, and the list is
+#' passed to an internal function \code{knitr:::wrap()} to return a character
+#' vector of final output.
 #' @param options a list of chunk options (usually this is just the object
 #'   \code{options} passed to the engine function; see
 #'   \code{\link{knit_engines}})
@@ -46,7 +53,14 @@ knit_engines = new_defaults()
 #' @return A character string generated from the source code and output using
 #'   the appropriate output hooks.
 #' @export
+#' @examples library(knitr)
+#' engine_output(opts_chunk$merge(list(engine = 'Rscript')), code = '1 + 1', out = '[1] 2')
+#' engine_output(opts_chunk$merge(list(echo = FALSE, engine = 'Rscript')), code = '1 + 1', out = '[1] 2')
+#'
+#' # expert use only
+#' engine_output(opts_chunk$merge(list(engine = 'python')), out = list(structure(list(src = '1 + 1'), class = 'source'), '2'))
 engine_output = function(options, code, out, extra = NULL) {
+  if (missing(code) && is.list(out)) return(unlist(wrap(out, options)))
   if (!is.logical(options$echo)) code = code[options$echo]
   if (length(code) != 1L) code = paste(code, collapse = '\n')
   if (options$engine == 'sas' && length(out) > 1L && !grepl('[[:alnum:]]', out[2]))